Fix #2 - Support node 0.8/0.10 at the same time
Before this fix, support for node 0.8 was on a different branch. This
commit gathers all the binaries under lib/bin/, and adds to code to
attempt to load all of them.

The code is farely straight forward. It gets a list of all the
dirs in bin/, then tries to load sqlserver.node from all of them. If
none of them was successfully loaded, an exception is thrown.

One thing that is not so great is having to keep all the binaries in
separate folders. Ideally all the binary files would be in bin/ and
have different filenames. For some reason, if `sqlserver.node` is
renamed to something else the require will fail. Maybe this can be
fixed somehow, but since I can't see an easy fix, I'll just use
different dirs for now.

Adding support for a new node version is just a matter of dropping
the binaries inside bin/ under their respective dirs.