Fix #5 - Reference binaries using abs paths
Relative pathnames with the `fs` module are not relative to the file
making the call (as intended in the code), but to the directory of
execution of the node process. From the node documentation:

    Relative path to filename can be used, remember however that this path
    will be relative to process.cwd().

This patch uses `__dirname` to construct an absolute path to `bin/`
based on the position of the `sqlserver.native.js` file. Bear in mind
that the `__dirname` object contains the dir of the file accessing
it.